Updated browser detection util (de-duped from playground)
