+++ weave-js/src/common/util/browser.ts
@@ -6,5 +6,12 @@ export const isFirefox = browser?.name === 'firefox';
 export const isSafari = browser?.name === 'safari';
 
 // navigator.platform is deprecated, so fallback to navigator.userAgent
-export const isMac =
-  navigator.platform?.startsWith('Mac') ?? navigator.userAgent.includes('Mac');
+export const isMac = () => {
+  const platform = navigator.platform || '';
+  const userAgent = navigator.userAgent || '';
+  const appVersion = navigator.appVersion || '';
+  const checkString = (str: string) => /Mac|iPhone|iPod|iPad/i.test(str);
+  return (
+    checkString(platform) || checkString(userAgent) || checkString(appVersion)
+  );
+};
