+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/PlaygroundPage/PlaygroundChat/PlaygroundChatInput.tsx
@@ -2,7 +2,7 @@ import {Box, Divider} from '@mui/material';
 import {MOON_250, MOON_500} from '@wandb/weave/common/css/color.styles';
 import {Button} from '@wandb/weave/components/Button';
 import React, {useState} from 'react';
-
+import {isMac} from '@wandb/weave/common/util/browser';
 import {StyledTextArea} from '../StyledTextarea';
 import {PlaygroundMessageRole} from '../types';
 
@@ -13,16 +13,6 @@ type PlaygroundChatInputProps = {
   onSend: (role: PlaygroundMessageRole, chatText: string) => void;
   onAdd: (role: PlaygroundMessageRole, chatText: string) => void;
   settingsTab: number | null;
-};
-
-const isMac = () => {
-  const platform = navigator.platform || '';
-  const userAgent = navigator.userAgent || '';
-  const appVersion = navigator.appVersion || '';
-  const checkString = (str: string) => /Mac|iPhone|iPod|iPad/i.test(str);
-  return (
-    checkString(platform) || checkString(userAgent) || checkString(appVersion)
-  );
 };
 
 export const PlaygroundChatInput: React.FC<PlaygroundChatInputProps> = ({
